feat: front-end, back de sales

+++ src/app/components/inventory/inventory.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { InventoryService } from '../../../services/inventory.service';
+import { SalesService } from '../../../services/sales.service';
 import { InventoryItem } from '../../../models/inventory.model';
+import { Sale } from '../../../models/sales.model';
 
 @Component({
   selector: 'app-inventory',
@@ -9,9 +11,12 @@ import { InventoryItem } from '../../../models/inventory.model';
 })
 export class InventoryComponent implements OnInit {
   items: InventoryItem[] = [];
-  searchTerm: string = ''; // Para a barra de pesquisa
+  searchTerm: string = '';
 
-  constructor(private inventoryService: InventoryService) { }
+  constructor(
+    private inventoryService: InventoryService,
+    private salesService: SalesService
+  ) { }
 
   ngOnInit(): void {
     this.loadItems();
@@ -27,17 +32,17 @@ export class InventoryComponent implements OnInit {
   }
 
   addItem(name: string, quantity: string, price: string): void {
-    const quantityNum = parseInt(quantity, 10); // Convertendo para número
-    const priceNum = parseFloat(price); // Convertendo para número
+    const quantityNum = parseInt(quantity, 10);
+    const priceNum = parseFloat(price);
 
-    // Verificar se o nome do item já existe
     const existingItem = this.items.find(item => item.name.toLowerCase() === name.toLowerCase());
     if (existingItem) {
       console.error('Produto com esse nome já existe');
-      return; // Impede a adição de itens duplicados
+      alert('Produto com esse nome já existe');
+      return;
     }
 
-    const newItem: InventoryItem = { name, quantity: quantityNum, price: priceNum }; // Removido o _id
+    const newItem: InventoryItem = { name, quantity: quantityNum, price: priceNum };
     this.inventoryService.addItem(newItem).subscribe(
       (item: InventoryItem) => {
         this.items.push(item);
@@ -52,23 +57,46 @@ export class InventoryComponent implements OnInit {
 
     this.inventoryService.updateItem(id, updatedItem).subscribe(
       () => {
-        // Atualiza o item no array sem sobrescrever _id
         this.items[index] = updatedItem;
       },
       (error) => console.error('Erro ao atualizar item', error)
     );
   }
 
-  deleteItem(id: string): void {
+  deleteItem(id: string | undefined): void {
+    if (!id) return;
+
     this.inventoryService.deleteItem(id).subscribe(
       () => {
+        console.log('Item deletado com sucesso');
+        this.salesService.getSalesByProductId(id).subscribe(
+          (sales: Sale[]) => {
+            if (sales.length > 0) {
+              sales.forEach((sale: Sale) => {
+                this.salesService.deleteSale(sale._id!).subscribe(
+                  () => {
+                    console.log('Venda deletada:', sale._id);
+                  },
+                  (error: any) => {
+                    console.error('Erro ao deletar venda', error);
+                  }
+                );
+              });
+            } else {
+              console.log('Nenhuma venda encontrada para deletar.');
+            }
+          },
+          (error: any) => {
+            console.error('Erro ao buscar vendas relacionadas', error);
+          }
+        );
+
         this.items = this.items.filter(item => item._id !== id);
       },
       (error) => console.error('Erro ao deletar item', error)
     );
   }
 
-  // Função para filtrar os itens com base na pesquisa
   get filteredItems(): InventoryItem[] {
     return this.items.filter(item =>
       item.name.toLowerCase().includes(this.searchTerm.toLowerCase())
